+++ tests/testthat/test-create-empty-textgrid.R
@@ -1,13 +1,12 @@
 s1 <- system.file("extdata", "test.wav", package = "phonfieldwork")
-tdir <- tempdir()
-dir.create(paste0(tdir, "/make_txtgrd"))
-file.copy(s1, paste0(tdir, "/make_txtgrd"))
-
-create_empty_textgrid(paste0(tdir, "/make_txtgrd/test.wav"))
-
+tdir <- tempdir(check = TRUE)
+create_empty_textgrid(get_sound_duration(s1),
+  path = tdir,
+  result_file_name = "test"
+)
 test_that("create_empty_textgrid", {
   expect_equal(
-    list.files(paste0(tdir, "/make_txtgrd"), "\\.TextGrid"),
+    list.files(tdir, "\\.TextGrid"),
     "test.TextGrid"
   )
 })